Refactor Modal functionality and update app name
Modal functionality was moved from the PostList component to the App component to improve cohesion and responsibility distribution. A boolean state value was also added to control whether the Modal is always visible on the page, enhancing the user's control. Lastly, the app's name was changed from "React Poster" to "React Tweet".