Add cancel button to new post form
A cancel button has been added to the new post form component. This button will be used to stop the creation of a new post if the user decides to do so. Its functionality was implemented in the PostList component and passed down as a prop to the NewPost component.