Refactor routes and components for posts
Moved "App.jsx" to "Posts.jsx" under the routes folder and made it the main route for the application. Also, the components "NewPost.jsx" was moved to routes and "Modal.jsx" was introduced to it. Additionally, elements for creating new posts were commented out in "PostList.jsx" to clean up unnecessary code.